split inverters into separate modules
