+++ quartz_solar_forecast/inverters/inverter_model.py
