+++ quartz_solar_forecast/inverters/solis/solis.py
@@ -10,12 +10,9 @@ import re
 from enum import Enum
 from http import HTTPStatus
 import json
-from typing import Any, Optional
+from typing import Any
 
-from pydantic import Field
-from pydantic_settings import BaseSettings
-
-from quartz_solar_forecast.inverters.inverter import AbstractInverter
+from quartz_solar_forecast.inverters.solis.solis_model import SolisSettings
 
 try:
     import async_timeout
@@ -32,24 +29,6 @@ VERB = "POST"
 INVERTER_LIST = RESOURCE_PREFIX + 'inverterList'
 INVERTER_DAY = RESOURCE_PREFIX + 'inverterDay'
 
-
-class SolisSettings(BaseSettings):
-    api_url: str = Field(alias="SOLIS_CLOUD_API_URL", default='https://www.soliscloud.com')
-    port: str = Field(alias="SOLIS_CLOUD_API_PORT", default='13333')
-    api_key: str = Field(alias="SOLIS_CLOUD_API_KEY")
-    client_secret: str = Field(alias="SOLIS_CLOUD_API_KEY_SECRET")
-
-
-class SolisInverter(AbstractInverter):
-    def __init__(self, settings: SolisSettings):
-        self.__settings = settings
-
-    def get_data(self, ts: pd.Timestamp) -> Optional[pd.DataFrame]:
-        try:
-            return asyncio.run(get_solis_data(self.__settings))
-        except Exception as e:
-            print(f"Error retrieving Solis data: {str(e)}")
-            return None
 
 class SoliscloudAPI():
     """Class with functions for reading data from the Soliscloud Portal."""
